refactor(loader): split validate.go into domain files and dedup run validation
Break the 1590-line validate.go into responsibility-scoped files (runner,
run, service, assert, artifact, store, fixture, pty, signal, cdp, http) and
unify the run-step validation that validateSuiteBlock and validateStep each
copied into a single validateRunStep helper (full=false for suite.setup,
full=true for scenario steps), preserving error text and ordering. Also
relocate two misplaced doc comments (validateDefaults, validateSSHRunFields)
to the declarations they document.
